LLMOps/End-to-end-RAG-using-Amazon-Bedrock/requirements.txt
langchain
langchain-community
langchain-aws
langchain-text-splitters
langchain-chroma
langchain-core
pypdf
streamlit
chromadb
boto3
python-dotenv
